Changed type of bit

+++ 0x14-bit_manipulation/4-clear_bit.c
@@ -8,7 +8,7 @@
 */
 int clear_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned int bit;
+	unsigned long int bit;
 
 	if (index > 63)
 		return (-1);
